moving App component to container dir
